new account version added, nonce storage optimized with hasher in lookup map
